Issue #58 refactor: 클래스 패지킹 변경
Item 클래스 -> model 패키지
adapter 패키지 삭제
